try 4 for task 8

+++ api/v1/views/cities.py
@@ -13,37 +13,51 @@ from models import storage
                  strict_slashes=False)
 def state_all_cities(state_id):
     """Retrieves all the cities of a given state_id"""
-    states = storage.all(State).values()
-    if state is None:
-        return abort(404)
-    for state in states:
+    for state in storage.all(State).values():
         if state.id == state_id:
             list_city = []
             for city in state.cities:
                 list_city.append(city.to_dict())
-    return jsonify(list_city)
-
-
-@app_views.route("/cities/<city_id>", methods=["GET"],
-                 strict_slashes=False)
-def retrieve_one_city(city_id):
-    """Retrieves one city of a given city_id"""
-    for city in storage.all(City).values():
-        if city.id == city_id:
-            return jsonify(city.to_dict())
+            return jsonify(list_city)
     return abort(404)
 
 
-@app_views.route('/cities/<city_id>', methods=['DELETE'],
+@app_views.route("/cities/<city_id>", methods=['GET', 'DELETE', 'PUT'],
                  strict_slashes=False)
-def delete_one_city(city_id):
-    """ Method that deletes a city based on its ID. """
-    for city in storage.all(City).values():
-        if city.id == city_id:
-            city.delete()
-            storage.save()
-            return jsonify({})
-    return abort(404)
+def cities_with_id(city_id):
+    """Methods that retrieves all methods for cities with id"""
+
+    if request.method == 'GET':
+        """Retrieves a city of a given city_id"""
+        for city in storage.all(City).values():
+            if city.id == city_id:
+                return jsonify(city.to_dict())
+        return abort(404)
+
+    if request.method == 'DELETE':
+        """Deletes a city of a given city_id """
+        for city in storage.all(City).values():
+            if city.id == city_id:
+                city.delete()
+                storage.save()
+                return jsonify({})
+        return abort(404)
+
+    if request.method == 'PUT':
+        """Updates a city of a given city_id"""
+        city = storage.get(City, city_id)
+        if city is None:
+            return abort(404)
+
+        r = request.get_json()
+        if r is None:
+            return abort(400, 'Not a JSON')
+        toIgnore = ["id", "created_at", "updated_it"]
+        for key, value in r.items():
+            if value not in toIgnore:
+                setattr(city, key, value)
+        city.save()
+        return jsonify(city.to_dict())
 
 
 @app_views.route("/states/<state_id>/cities", methods=["POST"],
@@ -63,22 +77,3 @@ def create_one_city(state_id):
     new = City(**r)
     new.save()
     return jsonify(new.to_dict()), 201
-
-
-@app_views.route('/cities/<city_id>', methods=['PUT'],
-                 strict_slashes=False)
-def put_city(city_id):
-    """Updates a city based on its id"""
-    city = storage.get(City, city_id)
-    if city is None:
-        return abort(404)
-
-    r = request.get_json()
-    if r is None:
-        return abort(400, 'Not a JSON')
-    toIgnore = ["id", "created_at", "updated_it"]
-    for key, value in r.items():
-        if value not in toIgnore:
-            setattr(city, key, value)
-    city.save()
-    return jsonify(city.to_dict())
